Detriplicate internal SingleProducerSingleConsumerQueue
System.Private.CoreLib, System.Threading.Tasks.Dataflow, and System.Threading.Channels have all ended up with their own copy of SingleProducerSingleConsumerQueue and its associated helpers.  This consolidates them down to a single shared copy.

There's no functional change here, just deleting duplicates and moving things around.
